refactor: :recycle: Code cleanup
Code cleanup according typescript & svelte features and rules
