+++ src/background.ts
@@ -1,9 +1,9 @@
-import type { IOptions } from "./types";
+import * as types from "./types";
 
 chrome.runtime.onInstalled.addListener(() => {
-    chrome.storage.sync.get({ count: 0 } as IOptions, ({ count }: IOptions) => {
+    /* chrome.storage.sync.get({ count: 0 } as IOptions, ({ count }: IOptions) => {
         console.log(count);
-    });
+    }); */
 });
 
 chrome.tabs.onUpdated.addListener((_, changed, tab) => {
