+++ src/popup.ts
@@ -1,4 +1,4 @@
-import Options from "./components/Options.svelte";
+import Options from "./components/settings/Options.svelte";
 import type { IOptions } from "./types";
 import App from "./App.svelte";
 
@@ -6,9 +6,9 @@ const app = new App({
     target: document.body
 });
 
-chrome.storage.sync.get({ count: 0 } as IOptions, ({ count }: IOptions) => {
-    const options = new Options({
-        target: document.body,
-        props: { count },
-    });
-});
+// chrome.storage.sync.get({ count: 0 } as Options, ({ count }: Options) => {
+//     const options = new Options({
+//         target: document.body,
+//         props: { count },
+//     });
+// });
