+++ src/types.ts
@@ -1,3 +1,12 @@
-export type IOptions = {
-    count: number;
-};
+export interface IItem {
+    label: string;
+    value: string;
+    icon: string;
+    description: string;
+}
+
+export interface IOptions {
+    labels: IItem[];
+    decorations: IItem[];
+
+}
